Update patch set 6
Patch Set 6:

With patchset 6, I am able to deploy DCAE plus a minimal subset of the rest of ONAP.  The dmaap-bc-dmaap-provision job runs successfully with this dbc-client.  The DFC also comes up, but in my
environment there's a missing unauthenticated MR topic that keeps the container from becoming ready.  The configuration, however, is correctly updated with the publish URL, publisher ID, etc.
Once DCAE was deployed, I deployed a standalone microservice configured with two feeds and two
publishers.  Both feeds and publishers were created, and the microservice's configuration was
correctly populated with publish URLs, publisher IDs, etc. 

(I've done no testing to verify that it's actually possible to publish to the feeds that are created, but it appears that the feeds and publishers are properly set up in DMaaP and the information is
correctly made available to the microservice through the applicationConfig if the proper environment variables are used.)

Patch-set: 6
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>